[router] Fix missing subpath warning from Metro when importing from `expo-router/server`
