+++ packages/expo-router/package.json
@@ -174,6 +174,14 @@
       "expo-source": "./src/layouts/TopTabs.tsx",
       "default": "./build/layouts/TopTabs.js"
     },
+    "./server": {
+      "types": {
+        "expo-source": "./src/server/index.ts",
+        "default": "./build/server/index.d.ts"
+      },
+      "expo-source": "./src/server/index.ts",
+      "default": "./build/server/index.js"
+    },
     "./stack": {
       "types": {
         "expo-source": "./src/stack.ts",
