+++ packages/expo-router/server.d.ts
@@ -1,14 +1,1 @@
-export type {
-  ImmutableRequest,
-  GenerateMetadataFunction,
-  LoaderFunction,
-  Metadata,
-  MiddlewareFunction,
-} from 'expo-server';
-
-export { createStaticLoader, createServerLoader } from 'expo-server';
-
-export type RequestHandler = (
-  request: Request,
-  params: Record<string, string>
-) => Response | Promise<Response>;
+export * from './build/server';
