+++ packages/expo-router/src/server/index.ts
@@ -1,0 +1,14 @@
+export type {
+  ImmutableRequest,
+  GenerateMetadataFunction,
+  LoaderFunction,
+  Metadata,
+  MiddlewareFunction,
+} from 'expo-server';
+
+export { createStaticLoader, createServerLoader } from 'expo-server';
+
+export type RequestHandler = (
+  request: Request,
+  params: Record<string, string>
+) => Response | Promise<Response>;
